+++ components/bridge/dao/api/src/main/java/com/flipperdevices/bridge/dao/api/delegates/key/SimpleKeyApi.kt
@@ -17,8 +17,6 @@ interface SimpleKeyApi {
 
     suspend fun insertKey(key: FlipperKey)
 
-    suspend fun updateKey(oldKey: FlipperKey, newKey: FlipperKey)
-
     suspend fun updateKeyContent(keyPath: FlipperKeyPath, content: FlipperKeyContent)
 
     suspend fun getKey(keyPath: FlipperKeyPath): FlipperKey?