+++ components/share/receive/src/main/java/com/flipperdevices/share/receive/composable/ComposableKeyReceive.kt
@@ -6,7 +6,6 @@ import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.flipperdevices.keyedit.api.KeyEditApi
 import com.flipperdevices.keyscreen.api.KeyScreenApi
 import com.flipperdevices.share.receive.model.ReceiveState
 import com.flipperdevices.share.receive.viewmodels.KeyReceiveViewModel
@@ -16,7 +15,6 @@ fun ComposableKeyReceive(
     keyScreenApi: KeyScreenApi,
     state: ReceiveState,
     viewModel: KeyReceiveViewModel,
-    editApi: KeyEditApi,
     onCancel: () -> Unit = {}
 ) {
     when (state) {
@@ -26,21 +24,14 @@ fun ComposableKeyReceive(
             keyParsed = state.parsed,
             savingInProgress = false,
             onSave = viewModel::onSave,
-            onCancel = onCancel,
-            onEdit = viewModel::onEdit
+            onCancel = onCancel
         )
         is ReceiveState.Saving -> ComposableKeySaveScreen(
             keyScreenApi = keyScreenApi,
             keyParsed = state.parsed,
             savingInProgress = true,
             onSave = viewModel::onSave,
-            onCancel = onCancel,
-            onEdit = viewModel::onEdit
-        )
-        is ReceiveState.Editing -> editApi.EditScreen(
-            flipperKey = state.flipperKey,
-            parsedKey = state.parsed,
-            onKeyEditFinished = viewModel::onCloseEdit
+            onCancel = onCancel
         )
         ReceiveState.Finished -> return
     }