+++ components/share/receive/src/main/java/com/flipperdevices/share/receive/model/ReceiveState.kt
@@ -16,10 +16,5 @@ sealed class ReceiveState {
         val parsed: FlipperKeyParsed
     ) : ReceiveState()
 
-    data class Editing(
-        val flipperKey: FlipperKey,
-        val parsed: FlipperKeyParsed
-    ) : ReceiveState()
-
     object Finished : ReceiveState()
 }